+++ tests/FIXTURES/enumdeps/expected.sql
@@ -2,9 +2,7 @@ drop view if exists "public"."v";
 
 drop view if exists "public"."v2";
 
-alter type "public"."e" rename to "e__old_version_to_be_dropped";
-
-create type "public"."e" as enum ('a', 'b', 'c', 'd');
+alter type "public"."e" add value 'd' after 'c';
 
 create table "public"."created_with_e" (
     "id" integer,
@@ -13,8 +11,6 @@ create table "public"."created_with_e" (
 
 
 alter table "public"."t" alter column category type "public"."e" using category::text::"public"."e";
-
-drop type "public"."e__old_version_to_be_dropped";
 
 create or replace view "public"."v" as  SELECT id,
     category